\docType{package}
\name{DiscriMiner-package}
\alias{DiscriMiner-package}
\title{Tools of the Trade for Discriminant Analysis}
\description{
  DiscriMiner contains several functions for Discriminant
  Analysis and Classification purposes covering various
  methods such as descriptive, geometric, linear,
  quadratic, PLS, as well as qualitative discriminant
  analyses.
}
\details{
  \tabular{ll}{ Package: \tab DiscriMiner\cr Type: \tab
  Package\cr Version: \tab 0.1-23\cr Date: \tab
  2012-12-20\cr License: \tab GPL-3\cr }
}
\author{
  Gaston Sanchez

  Maintainer: Gaston Sanchez <gaston.stat@gmail.com>
}
\references{
  \url{http://www.gastonsanchez.com/discriminer}

  Lebart L., Piron M., Morineau A. (2006) \emph{Statistique
  exploratoire multidimensionnelle}. Dunod, Paris.

  Nakache J-P., Confais J. (2003) \emph{Statistique
  explicative appliquee}. Editions Technip, Paris.

  Saporta G. (2006) \emph{Probabilites, analyse des donnees
  et statistique}. Editions Technip, Paris.

  Tenenhaus M. (1998) \emph{La Regression PLS}. Editions
  Technip, Paris.

  Tenenhaus M. (2007) \emph{Statistique}. Dunod, Paris.

  Tuffery S. (2008) \emph{Data Mining et Statistique
  Decisionnelle}. Editions Technip, Paris.

  Tuffery S. (2011) \emph{Data Mining and Statistics for
  Decision Making}. Wiley, Chichester.

  \emph{Multiple Correspondence Analysis and Related
  Methods}. (2006) Edited by Michael Greenacre and Jorg
  Blasius. Chapman and Hall/CRC
}
\keyword{package}

