\name{test_bad_learn_set}
\alias{test_bad_learn_set}
\title{Test Bad Learning Set in Disqual}
\usage{
  test_bad_learn_set(learn_data)
}
\arguments{
  \item{learn_data}{learning dataset}
}
\value{
  whether the learn data has constant values
}
\description{
  Test Bad Learning Set in Disqual
}
\keyword{internal}

